--- lightstep/resource_metric_dashboard_test.go
@@ -1,181 +1,0 @@
-package lightstep
-
-import (
-	"context"
-	"fmt"
-	"regexp"
-	"testing"
-
-	"github.com/lightstep/terraform-provider-lightstep/client"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
-)
-
-func TestAccMetricDashboard(t *testing.T) {
-	var dashboard client.UnifiedDashboard
-
-	uqlQuery := `metric pagerduty.task.success | filter (kube_app == "pagerduty") | rate | group_by ["cluster-name"], max`
-
-	// missing required field 'type'
-	badDashboard := fmt.Sprintf(`
-resource "lightstep_dashboard" "test" {
-  project_name          = "terraform-provider-tests"
-  dashboard_name        = "Acceptance Test Dashboard"
-  dashboard_description = "Dashboard to test if the terraform provider works"
-
-  chart {
-    name = "Chart Number One"
-    rank = 1
-
-    y_axis {
-      min = 0.4
-      max = 5.0
-    }
-
-
-    query {
-      hidden              = false
-      query_name          = "a"
-      display             = "bar"
-      query_string        = <<EOT
-%s
-EOT
-    }
-  }
-}
-`, uqlQuery)
-
-	dashboardConfig := fmt.Sprintf(`
-resource "lightstep_dashboard" "test" {
-  project_name          = "terraform-provider-tests"
-  dashboard_name        = "Acceptance Test Dashboard"
-  dashboard_description = "Dashboard to test if the terraform provider works"
-
-  chart {
-    name = "Chart Number One"
-    rank = 1
-    type = "timeseries"
-
-    y_axis {
-      min = 0.4
-      max = 5.0
-    }
-
-
-    query {
-      hidden              = false
-      query_name          = "a"
-      display             = "bar"
-      query_string        = <<EOT
-%s
-EOT
-    }
-  }
-}
-`, uqlQuery)
-
-	updatedTitleDashboardConfig := fmt.Sprintf(`
-resource "lightstep_dashboard" "test" {
-  project_name          = "terraform-provider-tests"
-  dashboard_name        = "Acceptance Test Dashboard Updated"
-  dashboard_description = "Dashboard to test if the terraform provider still works"
-
-  chart {
-    name = "Chart Number One"
-    rank = 1
-    type = "timeseries"
-
-    y_axis {
-      min = 0.4
-      max = 5.0
-    }
-
-
-    query {
-      hidden              = false
-      query_name          = "a"
-      display             = "bar"
-      query_string        = <<EOT
-%s
-EOT
-    }
-  }
-}
-`, uqlQuery)
-
-	resourceName := "lightstep_dashboard.test"
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:          func() { testAccPreCheck(t) },
-		ProviderFactories: testAccProviderFactories,
-		CheckDestroy:      testGetMetricDashboardDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: badDashboard,
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckMetricDashboardExists(resourceName, &dashboard)),
-				ExpectError: regexp.MustCompile("The argument \"type\" is required, but no definition was found."),
-			},
-			{
-				Config: dashboardConfig,
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckMetricDashboardExists(resourceName, &dashboard),
-					resource.TestCheckResourceAttr(resourceName, "dashboard_name", "Acceptance Test Dashboard"),
-					resource.TestCheckResourceAttr(resourceName, "dashboard_description", "Dashboard to test if the terraform provider works"),
-					resource.TestCheckResourceAttr(resourceName, "chart.0.name", "Chart Number One"),
-					resource.TestCheckResourceAttr(resourceName, "chart.0.rank", "1"),
-					resource.TestCheckResourceAttr(resourceName, "chart.0.type", "timeseries"),
-					resource.TestCheckResourceAttr(resourceName, "chart.0.query.0.query_string", uqlQuery+"\n"),
-				),
-			},
-			{
-				Config: updatedTitleDashboardConfig,
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckMetricDashboardExists(resourceName, &dashboard),
-					resource.TestCheckResourceAttr(resourceName, "dashboard_name", "Acceptance Test Dashboard Updated"),
-					resource.TestCheckResourceAttr(resourceName, "dashboard_description", "Dashboard to test if the terraform provider still works"),
-				),
-			},
-		},
-	})
-}
-
-func testGetMetricDashboardDestroy(s *terraform.State) error {
-	conn := testAccProvider.Meta().(*client.Client)
-	for _, r := range s.RootModule().Resources {
-		if r.Type != "metric_alert" {
-			continue
-		}
-
-		s, err := conn.GetUnifiedDashboard(context.Background(), test_project, r.Primary.ID, false)
-		if err == nil {
-			if s.ID == r.Primary.ID {
-				return fmt.Errorf("metric dashboard with ID (%v) still exists.", r.Primary.ID)
-			}
-		}
-	}
-	return nil
-}
-
-func testAccCheckMetricDashboardExists(resourceName string, dashboard *client.UnifiedDashboard) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		tfDashboard, ok := s.RootModule().Resources[resourceName]
-		if !ok {
-			return fmt.Errorf("not found: %s", resourceName)
-		}
-
-		if tfDashboard.Primary.ID == "" {
-			return fmt.Errorf("id is not set")
-		}
-
-		c := testAccProvider.Meta().(*client.Client)
-		dash, err := c.GetUnifiedDashboard(context.Background(), test_project, tfDashboard.Primary.ID, false)
-		if err != nil {
-			return err
-		}
-
-		dashboard = dash
-		return nil
-	}
-}
